+++ test/rtk_batch_test.dart
@@ -56,7 +56,6 @@ void main() {
       final batch = RTKBatch(
         context: const RTKContext(
           platform: 'ios',
-          environment: 'app_store',
           appVersion: '1.0.0',
           buildNumber: '100',
         ),
@@ -74,7 +73,6 @@ void main() {
       expect(json, {
         'context': {
           'platform': 'ios',
-          'environment': 'app_store',
           'app_version': '1.0.0',
           'build_number': '100',
         },