+++ test/rtk_test.dart
@@ -10,7 +10,6 @@ void main() {
   RTKConfig config() => RTKConfig(
         endpoint: Uri.parse('https://rena.example.com'),
         publicWriteKey: 'public_test',
-        environment: 'app_store',
         debug: false,
       );